Don't delete finished task and use Ready condition
