+++ pkg/operator/taskcontroller/controller.go
@@ -4,20 +4,17 @@ import (
 	"context"
 	"fmt"
 	"path/filepath"
-	"strconv"
 
 	"github.com/ccremer/clustercode/pkg/api/conditions"
 	"github.com/ccremer/clustercode/pkg/api/v1alpha1"
 	"github.com/ccremer/clustercode/pkg/internal/pipe"
 	internaltypes "github.com/ccremer/clustercode/pkg/internal/types"
-	"github.com/ccremer/clustercode/pkg/internal/utils"
 	pipeline "github.com/ccremer/go-command-pipeline"
 	"github.com/go-logr/logr"
 	batchv1 "k8s.io/api/batch/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
@@ -60,14 +57,15 @@ func (r *TaskReconciler) Provision(ctx context.Context, obj *v1alpha1.Task) (rec
 	p := pipeline.NewPipeline[*TaskContext]().WithBeforeHooks(pipe.DebugLogger[*TaskContext](tc), tc.resolver.Record)
 	p.WithSteps(
 		p.When(r.hasNoSlicesPlanned, "create split job", r.createSplitJob),
-		p.When(r.isSplitComplete, "ensure count job", r.ensureCountJob),
+		p.When(r.hasCondition(conditions.SplitComplete()), "ensure count job", r.ensureCountJob),
 		p.When(r.allSlicesFinished, "create merge job", r.createMergeJob),
 		p.WithNestedSteps("schedule next slice job", r.partialSlicesFinished,
 			p.NewStep("determine next slice index", r.determineNextSliceIndex),
 			p.When(r.nextSliceDetermined, "create slice job", r.createSliceJob),
-			p.When(r.nextSliceDetermined, "update status", r.updateStatus),
+			p.When(r.nextSliceDetermined, "update status", r.updateStatusWithProgressing),
 		),
-		p.When(r.isMergeFinished, "create cleanup job", r.ensureCleanupJob),
+		p.When(r.hasCondition(conditions.MergeComplete()), "create cleanup job", r.ensureCleanupJob),
+		p.When(r.hasCondition(conditions.Ready()), "cleanup jobs", r.cleanupJobs),
 	)
 	err := p.RunWithContext(tc)
 	return reconcile.Result{}, err
@@ -83,7 +81,8 @@ func (r *TaskReconciler) hasNoSlicesPlanned(ctx *TaskContext) bool {
 
 func (r *TaskReconciler) allSlicesFinished(ctx *TaskContext) bool {
 	cond := meta.FindStatusCondition(ctx.task.Status.Conditions, conditions.Progressing().Type)
-	if cond != nil {
+	if cond != nil && cond.Status == metav1.ConditionFalse {
+		// "progressing" is false once all slices are completed.
 		return len(ctx.task.Status.SlicesFinished) >= ctx.task.Spec.SlicesPlannedCount
 	}
 	return false
@@ -97,20 +96,10 @@ func (r *TaskReconciler) nextSliceDetermined(ctx *TaskContext) bool {
 	return ctx.nextSliceIndex >= 0
 }
 
-func (r *TaskReconciler) isSplitComplete(ctx *TaskContext) bool {
-	cond := meta.FindStatusCondition(ctx.task.Status.Conditions, conditions.SplitComplete().Type)
-	if cond != nil {
-		return cond.Status == metav1.ConditionTrue
+func (r *TaskReconciler) hasCondition(condition metav1.Condition) func(*TaskContext) bool {
+	return func(ctx *TaskContext) bool {
+		return meta.IsStatusConditionTrue(ctx.task.Status.Conditions, condition.Type)
 	}
-	return false
-}
-
-func (r *TaskReconciler) isMergeFinished(ctx *TaskContext) bool {
-	cond := meta.FindStatusCondition(ctx.task.Status.Conditions, conditions.MergeComplete().Type)
-	if cond != nil {
-		return cond.Status == metav1.ConditionTrue
-	}
-	return false
 }
 
 func (r *TaskReconciler) determineNextSliceIndex(ctx *TaskContext) error {
@@ -144,60 +133,7 @@ func (r *TaskReconciler) determineNextSliceIndex(ctx *TaskContext) error {
 	return nil
 }
 
-func (r *TaskReconciler) createSplitJob(ctx *TaskContext) error {
-	sourceMountRoot := filepath.Join("/clustercode", internaltypes.SourceSubMountPath)
-	intermediateMountRoot := filepath.Join("/clustercode", internaltypes.IntermediateSubMountPath)
-	variables := map[string]string{
-		"${INPUT}":      filepath.Join(sourceMountRoot, ctx.task.Spec.SourceUrl.GetPath()),
-		"${OUTPUT}":     getSegmentFileNameTemplatePath(ctx, intermediateMountRoot),
-		"${SLICE_SIZE}": strconv.Itoa(ctx.task.Spec.EncodeSpec.SliceSize),
-	}
-	job := &batchv1.Job{ObjectMeta: metav1.ObjectMeta{
-		Name:      fmt.Sprintf("%s-%s", ctx.task.Spec.TaskId, internaltypes.JobTypeSplit),
-		Namespace: ctx.task.Namespace,
-	}}
-
-	_, err := controllerutil.CreateOrUpdate(ctx, r.Client, job, func() error {
-		createFfmpegJobDefinition(job, ctx.task, &TaskOpts{
-			args:              utils.MergeArgsAndReplaceVariables(variables, ctx.task.Spec.EncodeSpec.DefaultCommandArgs, ctx.task.Spec.EncodeSpec.SplitCommandArgs),
-			jobType:           internaltypes.JobTypeSplit,
-			mountSource:       true,
-			mountIntermediate: true,
-		})
-		return controllerutil.SetControllerReference(ctx.task, job, r.Client.Scheme())
-	})
-	return err
-}
-
-func (r *TaskReconciler) createSliceJob(ctx *TaskContext) error {
-	ctx.resolver.MustRequireDependencyByFuncName(r.determineNextSliceIndex)
-
-	intermediateMountRoot := filepath.Join("/clustercode", internaltypes.IntermediateSubMountPath)
-	index := ctx.nextSliceIndex
-	variables := map[string]string{
-		"${INPUT}":  getSourceSegmentFileNameIndexPath(ctx, intermediateMountRoot, index),
-		"${OUTPUT}": getTargetSegmentFileNameIndexPath(ctx, intermediateMountRoot, index),
-	}
-	job := &batchv1.Job{ObjectMeta: metav1.ObjectMeta{
-		Name:      fmt.Sprintf("%s-%s-%d", ctx.task.Spec.TaskId, internaltypes.JobTypeSlice, index),
-		Namespace: ctx.task.Namespace,
-	}}
-	_, err := controllerutil.CreateOrUpdate(ctx, r.Client, job, func() error {
-		createFfmpegJobDefinition(job, ctx.task, &TaskOpts{
-			args:              utils.MergeArgsAndReplaceVariables(variables, ctx.task.Spec.EncodeSpec.DefaultCommandArgs, ctx.task.Spec.EncodeSpec.TranscodeCommandArgs),
-			jobType:           internaltypes.JobTypeSlice,
-			mountIntermediate: true,
-		})
-		job.Labels[internaltypes.ClustercodeSliceIndexLabelKey] = strconv.Itoa(index)
-
-		return controllerutil.SetControllerReference(ctx.task, job, r.Client.Scheme())
-	})
-	ctx.job = job
-	return err
-
-}
-
-func (r *TaskReconciler) updateStatus(ctx *TaskContext) error {
+func (r *TaskReconciler) updateStatusWithProgressing(ctx *TaskContext) error {
 	ctx.resolver.MustRequireDependencyByFuncName(r.createSliceJob)
 
 	meta.SetStatusCondition(&ctx.task.Status.Conditions, conditions.Progressing())
@@ -208,29 +144,12 @@ func (r *TaskReconciler) updateStatus(ctx *TaskContext) error {
 	return r.Client.Status().Update(ctx, ctx.task)
 }
 
-func (r *TaskReconciler) createMergeJob(ctx *TaskContext) error {
-	configMountRoot := filepath.Join("/clustercode", internaltypes.ConfigSubMountPath)
-	targetMountRoot := filepath.Join("/clustercode", internaltypes.TargetSubMountPath)
-	variables := map[string]string{
-		"${INPUT}":  filepath.Join(configMountRoot, v1alpha1.ConfigMapFileName),
-		"${OUTPUT}": filepath.Join(targetMountRoot, ctx.task.Spec.TargetUrl.GetPath()),
-	}
-	job := &batchv1.Job{ObjectMeta: metav1.ObjectMeta{
-		Name:      fmt.Sprintf("%s-%s", ctx.task.Spec.TaskId, internaltypes.JobTypeMerge),
-		Namespace: ctx.task.Namespace,
-	}}
-
-	_, err := controllerutil.CreateOrUpdate(ctx, r.Client, job, func() error {
-		createFfmpegJobDefinition(job, ctx.task, &TaskOpts{
-			args:              utils.MergeArgsAndReplaceVariables(variables, ctx.task.Spec.EncodeSpec.DefaultCommandArgs, ctx.task.Spec.EncodeSpec.MergeCommandArgs),
-			jobType:           internaltypes.JobTypeMerge,
-			mountIntermediate: true,
-			mountTarget:       true,
-			mountConfig:       true,
-		})
-		return controllerutil.SetControllerReference(ctx.task, job, r.Client.Scheme())
-	})
-	return err
+func (r *TaskReconciler) cleanupJobs(ctx *TaskContext) error {
+	return r.Client.DeleteAllOf(ctx, &batchv1.Job{},
+		client.MatchingLabels(ctx.task.Spec.TaskId.AsLabels()),
+		client.InNamespace(ctx.task.Namespace),
+		client.PropagationPolicy(metav1.DeletePropagationBackground),
+	)
 }
 
 func getSegmentFileNameTemplatePath(ctx *TaskContext, intermediateMountRoot string) string {
